refactor: move all server code
